refactor: move TPCH CLI implementation into tpcgen
