+++ tpcgen/src/lib.rs
@@ -1,3 +1,3 @@
-mod tpcds_cli;
-
-pub mod tpcgen_cli;
+//! This library contains both the TPCH and TPCDS command line clients.
+pub mod tpcds_cli;
+pub mod tpch_cli;
