+++ tpchgen-cli/bin/tpchgen_cli.rs
@@ -1,6 +1,6 @@
 use clap::Parser;
 use std::io;
-use tpchgen_cli::tpch_cli::Cli;
+use tpcgen::tpch_cli::Cli;
 
 #[tokio::main]
 async fn main() -> io::Result<()> {
